% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data108}
\alias{data108}
\title{Data from experiment ALEXIS 108: Main experiment, lateralization}
\format{A data frame with 37,397 rows and 12 columns.
\describe{
  \item{id}{Participant unique id number = 1007}
  \item{date}{Date of testing}
  \item{time}{Time of day for each trial}
  \item{day}{Rank order of date the data was collected}
  \item{session}{Session number, there were 13 sessions, run on the same day}
  \item{trial}{Trial number within session, from 1 to whenever the staircase stopped}
  \item{tau}{Inter-click interval in ms, Note: -1 is for sessions with no lead click}
  \item{ild}{Inter-arual-level difference of variable, always 10 dB}
  \item{ratio}{dB-difference between lead and lag click of variable}
  \item{revno}{Reversal number, equals 0 of the trial was not a reversal}
  \item{correct}{Ild pointing left (1) or right (0)}
  \item{answer}{The answer given by the listener, "left" (1) or "right" (0).
                The answer was correct if answer == correct}
  }}
\usage{
data108
}
\description{
This is data from an experiment with one participant (male, 53 year old male).
The listener was tested totally 60 days, 13 sessions each day, to test if it
is possible to improve the ability to lateralize lag clicks based on inter-aural
determine i each on a separate day, measuring level (ILD) differences. The lag
clicks were preceded by a lead click with no ILD. The time separation of lead
and lag clicks varied between 0 and 18 ms. The lag-click always had an
inter-aural level difference of 10 dB and no inter-aural-time difference.
For details, see vignette.
}
\keyword{datasets}
