% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data108extra}
\alias{data108extra}
\title{Data from experiment ALEXIS 108: Extra experiment, detection}
\format{A data frame with 12,115 rows and 12 columns.
\describe{
  \item{id}{Participant unique id number = 1007}
  \item{date}{Date of testing}
  \item{time}{Time of day for each trial}
  \item{day}{Rank order of date the data was collected}
  \item{session}{Session number, there were 13 sessions, run on the same day}
  \item{trial}{Trial number within session, from 1 to whenever the staircase stopped}
  \item{tau}{Inter-click interval in ms, Note: -1 is for sessions with no lead click}
  \item{ild}{Inter-arual-level difference of variable, always 10 dB}
  \item{ratio}{dB-difference between lead and lag click of variable}
  \item{revno}{Reversal number, equals 0 of the trial was not a reversal}
  \item{correct}{Position of variable, in third interval (0) or second interval (1)}
  \item{answer}{The answer given by the listener, "third" (0) or "second" (1).
                The answer was correct if answer == correct}
  }}
\usage{
data108extra
}
\description{
This is data from an experiment with one participant (male, 53 year old male).
The listener was tested for 20 days, 12 sessions per day. The tests measured
thresholds for detecting a lag-click in click pairs separated by an
inter-click-interval. For details, see vignette.
}
\keyword{datasets}
