% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepost_stan}
\alias{prepost_stan}
\title{Processed data from experiment ALEXIS 108. For creating Figure 3.}
\format{An object of class \code{data.frame} with 13 rows and 16 columns.}
\usage{
prepost_stan
}
\description{
Threshold estimates derived from data108 using Bayesian inference and
Hamiltonian Monte Carlo.
}
\keyword{datasets}
