% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standata}
\alias{standata}
\title{Processed data from experiment ALEXIS 108. For creating Figure 2.}
\format{An object of class \code{data.frame} with 663 rows and 7 columns.}
\usage{
standata
}
\description{
Threshold estimates derived from data108 using Bayesian inference and
Hamiltonian Monte Carlo.
}
\keyword{datasets}
